+++ src/discljord/messaging.clj
@@ -68,11 +68,12 @@
                            prepend-major-var? (cons [(keyword major-var) major-var-type])
                            true vec)
         spec-keys (vec (map spec-for opts))
-        unqualified-params (cond->> (map (comp symbol name) params) prepend-major-var? (cons major-var))]
+        unqualified-params (cond->> (map (comp symbol name) params) prepend-major-var? (cons major-var))
+        unqualified-opts (mapv (comp keyword name) opts)]
     `(do
        (defn ~endpoint-name
          ~doc-str
-         [~'conn ~@unqualified-params ~'& {:keys ~opts :as ~'opts}]
+         [~'conn ~@unqualified-params ~'& {:keys ~unqualified-opts :as ~'opts}]
          (let [user-agent# (:user-agent ~'opts)
                audit-reason# (:audit-reason ~'opts)
                p# (util/derefable-promise-chan)
@@ -804,7 +805,7 @@
   [application-id guild-id commands]
   [])
 
-(defendpoint get-guild-guild-application-command-permissions! nil
+(defendpoint get-guild-application-command-permissions! nil
   "Returns a promise containing the permission settings for all application commands accessible from the guild."
   [application-id guild-id]
   [])
@@ -819,6 +820,14 @@
 
   Returns a promise containing the updated permission settings in a map with some additional information."
   [application-id guild-id command-id ms.command/permissions]
+  [])
+
+(defendpoint batch-edit-application-command-permissions! nil
+  "Batch edits the permission settings for all commands in a guild.
+
+  This will overwrite all existing permissions for all commands in the guild.
+  Returns a promise containing the updated permission settings."
+  [application-id guild-id ms.command.guild/permissions-array]
   [])
 
 ;; -------------------------------------------------