+++ src/discljord/messaging/specs.clj
@@ -243,7 +243,7 @@
 
 (s/def :command.option/type (set (vals command-option-types)))
 
-(s/def :command.option/name (string-spec #"[\w-]{1,32}"))
+(s/def :command.option/name (string-spec #"\S{1,32}"))
 (s/def :command.option/description (string-spec 1 100))
 (s/def :command.option/default boolean?)
 (s/def :command.option/required boolean?)
@@ -286,6 +286,16 @@
 
 (s/def :discljord.messaging.specs.command/permissions (s/coll-of :command/permission))
 
+(s/def ::command-id ::ds/snowflake)
+(s/def :command/id ::command-id)
+
+(s/def :discljord.messaging.specs.command.guild/permissions
+  (s/keys :req-un [:command/id
+                   :discljord.messaging.specs.command/permissions]))
+
+(s/def :discljord.messaging.specs.command.guild/permissions-array
+  (s/coll-of :discljord.messaging.specs.command.guild/permissions))
+
 (s/def :discljord.messaging.specs.command/options
   (s/and (s/coll-of :command/option)
          (comp (partial >= 25) count)
@@ -299,7 +309,7 @@
 
 (s/def :command.option/options :discljord.messaging.specs.command/options)
 
-(s/def :discljord.messaging.specs.command/name (string-spec #"[\w-]{1,32}"))
+(s/def :discljord.messaging.specs.command/name (string-spec #"\S{1,32}"))
 
 (s/def :discljord.messaging.specs.command/description (string-spec 1 100))
 
@@ -319,8 +329,6 @@
                4000))))
 
 (s/def ::commands (s/and (s/coll-of ::command) #(<= (count %) 100)))
-
-(s/def ::command-id ::ds/snowflake)
 
 (s/def ::interaction-id ::ds/snowflake)
 (s/def ::interaction-token string?)